+++ src/fs/diskd/store_dir_diskd.cc
@@ -1,6 +1,6 @@
 
 /*
- * $Id: store_dir_diskd.cc,v 1.56 2001/10/17 19:33:06 hno Exp $
+ * $Id: store_dir_diskd.cc,v 1.57 2001/10/24 07:45:38 hno Exp $
  *
  * DEBUG: section 47    Store Directory Routines
  * AUTHOR: Duane Wessels
@@ -636,8 +636,8 @@ storeDiskdDirRebuildFromDirectory(void *data)
 	} else if (tmpe.swap_file_sz == sb.st_size - swap_hdr_len) {
 	    tmpe.swap_file_sz = sb.st_size;
 	} else if (tmpe.swap_file_sz != sb.st_size) {
-	    debug(20, 1) ("storeDiskdDirRebuildFromDirectory: SIZE MISMATCH %d!=%d\n",
-		tmpe.swap_file_sz, (int) sb.st_size);
+	    debug(20, 1) ("storeDiskdDirRebuildFromDirectory: SIZE MISMATCH %ld!=%ld\n",
+		(long int) tmpe.swap_file_sz, (long int) sb.st_size);
 	    storeDiskdDirUnlinkFile(SD, filn);
 	    continue;
 	}
@@ -877,7 +877,7 @@ storeDiskdDirRebuildFromSwapLog(void *data)
 }
 
 static int
-storeDiskdDirGetNextFile(RebuildState * rb, sfileno *filn_p, int *size)
+storeDiskdDirGetNextFile(RebuildState * rb, sfileno * filn_p, int *size)
 {
     SwapDir *SD = rb->sd;
     diskdinfo_t *diskdinfo = SD->fsdata;
@@ -1863,8 +1863,8 @@ storeDiskdCleanupDoubleCheck(SwapDir * sd, StoreEntry * e)
 	debug(20, 0) ("storeDiskdCleanupDoubleCheck: FILENO %08X\n", e->swap_filen);
 	debug(20, 0) ("storeDiskdCleanupDoubleCheck: PATH %s\n",
 	    storeDiskdDirFullPath(sd, e->swap_filen, NULL));
-	debug(20, 0) ("storeDiskdCleanupDoubleCheck: ENTRY SIZE: %d, FILE SIZE: %d\n",
-	    e->swap_file_sz, (int) sb.st_size);
+	debug(20, 0) ("storeDiskdCleanupDoubleCheck: ENTRY SIZE: %ld, FILE SIZE: %ld\n",
+	    (long int) e->swap_file_sz, (long int) sb.st_size);
 	storeEntryDump(e, 0);
 	return -1;
     }
