+++ src/structs.h
@@ -1,6 +1,6 @@
 
 /*
- * $Id: structs.h,v 1.406 2001/10/19 22:34:49 hno Exp $
+ * $Id: structs.h,v 1.407 2001/10/24 07:45:35 hno Exp $
  *
  *
  * SQUID Web Proxy Cache          http://www.squid-cache.org/
@@ -2081,7 +2081,7 @@ struct _helper_stateful_server {
 	int deferbyfunc;
 	int deferbycb;
     } stats;
-    size_t deferred_requests;	/* current number of deferred requests */
+    int deferred_requests;	/* current number of deferred requests */
     void *data;			/* State data used by the calling routines */
 };
